Refactor route and container components
