+++ src/containers/LandingContainer.tsx
@@ -1,0 +1,10 @@
+const LandingContainer = () => {
+  return (
+    <div>
+      <h1>Welcome to the Landing Page</h1>
+      <p>This is the landing page of your application.</p>
+    </div>
+  );
+};
+
+export default LandingContainer;
