+++ src/containers/ProgressContainer.tsx
@@ -1,0 +1,10 @@
+const ProgressContainer = () => {
+  return (
+    <div>
+      <h1>Welcome to the Progress Page</h1>
+      <p>This is the Progress page of your application.</p>
+    </div>
+  );
+};
+
+export default ProgressContainer;
