+++ src/containers/SettingContainer.tsx
@@ -1,0 +1,10 @@
+const SettingContainer = () => {
+  return (
+    <div>
+      <h1>Welcome to the Setting Page</h1>
+      <p>This is the Setting page of your application.</p>
+    </div>
+  );
+};
+
+export default SettingContainer;
